pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "microstructure_orientation"
authors = [
    {name = "Antoine Weisrock", email = "antoine.weisrock@centralelille.fr"},
]
maintainers = [
    {name = "Antoine Weisrock", email = "antoine.weisrock@gmail.com"},
]
description ="""A python module for the analysis of microstructure orientation
of transparency images of soft tissue, and the use of a Kelvin-based
mechanical model to describe their mechanical behavior."""
readme = "README.md"
requires-python = ">=3.12,<3.13"
keywords = ["Microstructure", "Orientaiton", "Gabor filter"]
license-files = ["LICENSE.txt"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy",
    "tqdm",
    "opencv-python",
    "cucim-cu12",
    "cupy-cuda12x",
    "numba",
    "scipy"
]

version = "0.0.1"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false